Learning the kivy language, new files created
