+++ main.py
@@ -1,54 +1,26 @@
-from re import MULTILINE
 import kivy
 from kivy.app import App
 from kivy.uix.label import Label # Import the simbols and widgets
-import os
-#os.environ["SDL_VIDEODRIVER"] = "x11"
 from kivy.uix.gridlayout import GridLayout # Import GridLayout design
 from kivy.uix.textinput import TextInput
 from kivy.uix.button import Button # import buttons
+from kivy.uix.widget import Widget # import tu use kv language
+from kivy.properties import ObjectProperty #Import object properties, to get the values
 
-class MenuLayout(GridLayout):
-    def __init__(self, **kwargs): # handle as many kwargs as come
-        super(MenuLayout, self).__init__(**kwargs)
-        self.cols = 1
-        
-        self.inside = GridLayout() # first layout
-        self.inside.cols = 2
+class MyGrid(Widget):
+    # Get the variable values on the kv file
+    name = ObjectProperty(None) # Initialize name
+    email = ObjectProperty(None)
 
-        #self.cols = 2
-        self.inside.add_widget(Label(text = "First Name: "))
-        self.firstName = TextInput(multiline = False)
-        self.inside.add_widget(self.firstName)
-
-        self.inside.add_widget(Label(text = "Last Name: "))
-        self.lastName = TextInput(multiline = False)
-        self.inside.add_widget(self.lastName)
-
-        self.inside.add_widget(Label(text = "Email: "))
-        self.email = TextInput(multiline = False)
-        self.inside.add_widget(self.email)
-
-        self.add_widget(self.inside)
-
-
-        self.submit = Button(text = "Submit:", font_size = 40)
-        self.submit.bind(on_press = self.pressed)
-        self.add_widget(self.submit)
-
-    def pressed(self, instance):
-        name = self.firstName.text
-        last = self.lastName.text
-        email = self.email.text
-
-        print("Name: ", name, "LastName: ", last, "Email: ", email)
-        self.lastName.text = ""
-        self.firstName.text = ""
+    # Onpress on the kv file button
+    def btn(self):
+        print("Name: ", self.name.text, " Email: ", self.email.text)
+        self.name.text = ""
         self.email.text = ""
 
 class MyApp(App):
     def build(self): # Construir o UI
-        return MenuLayout()
+        return MyGrid()
 
 if __name__ == "__main__":
     MyApp().run()
